refactor: update component imports to use new 'sppg' feature structure
