+++ src/feature/sppg/detail/components/DetailSupplierOpenCart.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import KeranjangButton from "@/feature/supplier/detail/components/KeranjangButton";
-import { useDetailSupplierSectionContext } from "@/feature/supplier/detail/components/DetailSupplierSectionProvider";
+import KeranjangButton from "@/feature/sppg/detail/components/KeranjangButton";
+import { useDetailSupplierSectionContext } from "@/feature/sppg/detail/components/DetailSupplierSectionProvider";
 
 const DetailSupplierOpenCart = () => {
   const { setCartOpened, canOpenCart } = useDetailSupplierSectionContext();
