+++ src/feature/sppg/detail/components/DetailSupplierSectionProvider.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { createContext, type ReactNode } from "react";
-import { useDetailSupplierSection } from "@/feature/supplier/detail/hooks/use-detail-supplier-section";
+import { createContext, useContext, type ReactNode } from "react";
+import { useDetailSupplierSection } from "@/feature/sppg/detail/hooks/use-detail-supplier-section";
 
 type DetailSupplierSectionState = ReturnType<typeof useDetailSupplierSection>;
 
@@ -20,6 +20,18 @@ const DetailSupplierSectionProvider = ({ children }: Props) => {
       {children}
     </DetailSupplierSectionContext.Provider>
   );
+};
+
+export const useDetailSupplierSectionContext = (): DetailSupplierSectionState => {
+  const context = useContext(DetailSupplierSectionContext);
+
+  if (context === null) {
+    throw new Error(
+      "useDetailSupplierSectionContext must be used within DetailSupplierSectionProvider",
+    );
+  }
+
+  return context;
 };
 
 export default DetailSupplierSectionProvider;
