+++ src/feature/sppg/detail/components/DetailSupplierToggleProducts.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import TampilkanButton from "@/feature/supplier/detail/components/TampilkanButton";
-import { useDetailSupplierSectionContext } from "@/feature/supplier/detail/components/DetailSupplierSectionProvider";
+import TampilkanButton from "@/feature/sppg/detail/components/TampilkanButton";
+import { useDetailSupplierSectionContext } from "@/feature/sppg/detail/components/DetailSupplierSectionProvider";
 
 const DetailSupplierToggleProducts = () => {
   const { showAllProducts, setShowAllProducts } =
